- moves handlers to a dedicated namespace
