+++ src/kiota/Handlers/BaseKiotaCommandHandler.cs
@@ -6,7 +6,7 @@ using Kiota.Builder.Configuration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace kiota;
+namespace kiota.Handlers;
 
 internal abstract class BaseKiotaCommandHandler : ICommandHandler
 {
